Transferencia da tratativa do questao_aluno_tai do iniciar prova tai para o obter questão tai
